Fix compile error

+++ hoomd/hpmc/IntegratorHPMCMonoGPUDepletantsTypes.cuh
@@ -62,7 +62,7 @@ struct hpmc_implicit_args_t
 template< class Shape >
 void hpmc_insert_depletants(const hpmc_args_t& args, const hpmc_implicit_args_t& implicit_args, const typename Shape::param_type *params);
 
-void generate_num_depletants(const unsigned int seed,
+void generate_num_depletants(const uint16_t seed,
                              const uint64_t timestep,
                              const unsigned int select,
                              const unsigned int rank,
